rename section to sections
